Avoiding deleting tag and category if there are still posts

+++ app/Http/Controllers/CategoriesController.php
@@ -98,6 +98,11 @@ class CategoriesController extends Controller
      */
     public function destroy(Category $category)
     {
+        if ($category->posts->count() > 0) {
+            session()->flash('error','Posts not yet empty.Cannot delete category.');
+            return redirect()->back();
+        }
+        
         $category->delete();
 
         session()->flash('success', 'Category deleted successfully');
