+++ app/Http/Controllers/TagsController.php
@@ -80,7 +80,7 @@ class TagsController extends Controller
      */
     public function update(UpdateTagsRequest $request, Tag $tag)
     {
-        $Tag->update([
+        $tag->update([
             'name' => $request->name
         ]);
 
@@ -96,6 +96,11 @@ class TagsController extends Controller
      */
     public function destroy(Tag $tag)
     {
+        if ($tag->posts->count() > 0 ) {
+            session()->flash('error', 'Tag cannot be deleted because it is associated to some posts.');
+
+            return redirect()->back();
+        }
         $tag->delete();
 
         session()->flash('success', 'Tag deleted successfully');
